Move FakeAppboy into the test where it's used
